refactor(types): split godot-runner.ts helpers into focused modules
Relocate non-class exports from godot-runner.ts to dedicated modules so
consumers import each symbol from the module that owns it. Pure mechanical
move — no behavior change.

- src/mcp.types.ts             shared MCP contract types (OperationParams,
                               ToolDefinition, ToolResponse, ToolHandler)
- src/utils/output-parsing.ts  extractJson, cleanOutput, cleanStdout,
                               normalizeForCompare
- src/utils/path-validation.ts validatePath, validateSubPath, validateNodePath,
                               isUnderDir, projectGodotPath, checkDisplayAvailable
- src/utils/error-response.ts  createErrorResponse, getErrorMessage,
                               extractGdError, validateProjectArgs,
                               validateSceneArgs
- src/utils/parameter-conversion.ts  parameterMappings, normalizeParameters,
                                     convertCamelToSnakeCase

godot-runner.ts retains the GodotRunner class plus its own coupled types
(GodotProcess, RuntimeSessionMode, RuntimeStopResult, GodotServerConfig,
OperationResult, BridgeDisconnectedError, InFlightCommand, readBytesFromChunks)
and the BRIDGE_*_MS timing constants. Lands at 982 lines.

Updates every import site (no barrel re-export). Docs (CONTRIBUTING.md,
docs/architecture.md, docs/tool-authoring.md) updated to point at the new
module ownership.
