+++ src/mcp.types.ts
@@ -1,0 +1,40 @@
+import type { GodotRunner } from './utils/godot-runner.js';
+
+export interface OperationParams {
+  [key: string]: unknown;
+}
+
+interface ToolAnnotations {
+  readOnlyHint?: boolean;
+  destructiveHint?: boolean;
+  idempotentHint?: boolean;
+  openWorldHint?: boolean;
+  title?: string;
+}
+
+export interface ToolDefinition {
+  name: string;
+  description: string;
+  inputSchema: {
+    type: string;
+    properties: Record<string, unknown>;
+    required: string[];
+  };
+  outputSchema?: {
+    type: string;
+    properties?: Record<string, unknown>;
+    required?: string[];
+  };
+  annotations?: ToolAnnotations;
+}
+
+export interface ToolResponse {
+  content: Array<{ type: string; text?: string; [k: string]: unknown }>;
+  isError?: boolean;
+  [k: string]: unknown;
+}
+
+export type ToolHandler = (
+  runner: GodotRunner,
+  args: OperationParams,
+) => Promise<ToolResponse> | ToolResponse;
