+++ src/tools/node-tools.ts
@@ -1,13 +1,10 @@
 import { existsSync } from 'fs';
 import { join } from 'path';
-import type { GodotRunner, OperationParams, ToolDefinition } from '../utils/godot-runner.js';
-import {
-  normalizeParameters,
-  validateSubPath,
-  validateNodePath,
-  createErrorResponse,
-  validateSceneArgs,
-} from '../utils/godot-runner.js';
+import type { GodotRunner } from '../utils/godot-runner.js';
+import type { OperationParams, ToolDefinition } from '../mcp.types.js';
+import { normalizeParameters } from '../utils/parameter-conversion.js';
+import { validateSubPath, validateNodePath } from '../utils/path-validation.js';
+import { createErrorResponse, validateSceneArgs } from '../utils/error-response.js';
 import { executeSceneOp } from '../utils/handler-helpers.js';
 
 // --- Tool definitions ---
