+++ src/tools/scene-tools.ts
@@ -1,14 +1,14 @@
 import { join } from 'path';
 import { existsSync } from 'fs';
-import type { GodotRunner, OperationParams, ToolDefinition } from '../utils/godot-runner.js';
+import type { GodotRunner } from '../utils/godot-runner.js';
+import type { OperationParams, ToolDefinition } from '../mcp.types.js';
+import { normalizeParameters } from '../utils/parameter-conversion.js';
+import { validateSubPath, validateNodePath } from '../utils/path-validation.js';
 import {
-  normalizeParameters,
-  validateSubPath,
-  validateNodePath,
   createErrorResponse,
   validateProjectArgs,
   validateSceneArgs,
-} from '../utils/godot-runner.js';
+} from '../utils/error-response.js';
 import { executeSceneOp } from '../utils/handler-helpers.js';
 
 export const sceneToolDefinitions: ToolDefinition[] = [
