+++ src/utils/godot-runner.ts
@@ -1,5 +1,5 @@
 import { fileURLToPath } from 'url';
-import { join, dirname, normalize, resolve, sep } from 'path';
+import { join, dirname, normalize } from 'path';
 import { existsSync } from 'fs';
 import type { ChildProcess, SpawnOptions } from 'child_process';
 import { spawn } from 'child_process';
@@ -15,6 +15,10 @@ import {
   MAX_FRAME_BYTES,
 } from './bridge-protocol.js';
 import { logDebug, logError, DEBUG_MODE } from './logger.js';
+import type { OperationParams } from '../mcp.types.js';
+import { cleanStdout, normalizeForCompare } from './output-parsing.js';
+import { checkDisplayAvailable, validateSubPath } from './path-validation.js';
+import { convertCamelToSnakeCase } from './parameter-conversion.js';
 
 /**
  * Thrown when the bridge socket closes (Godot exited, port closed, or peer
@@ -43,89 +47,6 @@ const BRIDGE_SHUTDOWN_ATTACHED_TIMEOUT_MS = 1500;
 const BRIDGE_PROCESS_EXIT_TIMEOUT_MS = 2000;
 const BRIDGE_RECONNECT_DELAY_MS = 1000;
 
-/**
- * Normalize a path for cross-platform comparison.
- * Folds Windows backslashes to forward slashes and strips trailing slashes,
- * so Node's `path.normalize` output matches Godot's `globalize_path("res://")`.
- */
-export function normalizeForCompare(p: string): string {
-  return normalize(p).replace(/\\/g, '/').replace(/\/+$/, '');
-}
-
-/**
- * Extract JSON from Godot output by finding the first { or [ and matching to the end.
- * This strips debug logs, version banners, and other noise.
- */
-export function extractJson(output: string): string {
-  // Find the first occurrence of { or [
-  const jsonStartBrace = output.indexOf('{');
-  const jsonStartBracket = output.indexOf('[');
-
-  let jsonStart = -1;
-  if (jsonStartBrace === -1 && jsonStartBracket === -1) {
-    return output; // No JSON found, return as-is
-  } else if (jsonStartBrace === -1) {
-    jsonStart = jsonStartBracket;
-  } else if (jsonStartBracket === -1) {
-    jsonStart = jsonStartBrace;
-  } else {
-    jsonStart = Math.min(jsonStartBrace, jsonStartBracket);
-  }
-
-  // Extract from JSON start to end
-  const jsonPart = output.substring(jsonStart);
-
-  // Try to parse to validate, if it fails return original
-  try {
-    JSON.parse(jsonPart.trim());
-    return jsonPart.trim();
-  } catch {
-    // If the extracted part isn't valid JSON, try to find the last } or ]
-    const lastBrace = jsonPart.lastIndexOf('}');
-    const lastBracket = jsonPart.lastIndexOf(']');
-    const lastEnd = Math.max(lastBrace, lastBracket);
-
-    if (lastEnd > 0) {
-      const extracted = jsonPart.substring(0, lastEnd + 1);
-      try {
-        JSON.parse(extracted);
-        return extracted;
-      } catch {
-        return output; // Return original if still can't parse
-      }
-    }
-    return output;
-  }
-}
-
-/**
- * Strip Godot banner and debug lines from output, keeping only meaningful content.
- */
-export function cleanOutput(output: string): string {
-  const lines = output.split('\n');
-  const cleanedLines = lines.filter((line) => {
-    const trimmed = line.trim();
-    // Skip empty lines
-    if (!trimmed) return false;
-    // Skip Godot version banner
-    if (trimmed.startsWith('Godot Engine v')) return false;
-    // Skip debug lines
-    if (trimmed.startsWith('[DEBUG]')) return false;
-    // Skip info lines that are just status updates
-    if (trimmed.startsWith('[INFO] Operation:')) return false;
-    if (trimmed.startsWith('[INFO] Executing operation:')) return false;
-    return true;
-  });
-  return cleanedLines.join('\n');
-}
-
-export function cleanStdout(stdout: string): string {
-  if (stdout.includes('{') || stdout.includes('[')) {
-    return extractJson(stdout);
-  }
-  return cleanOutput(stdout);
-}
-
 export interface GodotProcess {
   process: ChildProcess;
   output: string[];
@@ -150,349 +71,9 @@ export interface GodotServerConfig {
   debugMode?: boolean;
 }
 
-export interface OperationParams {
-  [key: string]: unknown;
-}
-
 export interface OperationResult {
   stdout: string;
   stderr: string;
-}
-
-export interface ToolAnnotations {
-  readOnlyHint?: boolean;
-  destructiveHint?: boolean;
-  idempotentHint?: boolean;
-  openWorldHint?: boolean;
-  title?: string;
-}
-
-export interface ToolDefinition {
-  name: string;
-  description: string;
-  inputSchema: {
-    type: string;
-    properties: Record<string, unknown>;
-    required: string[];
-  };
-  outputSchema?: {
-    type: string;
-    properties?: Record<string, unknown>;
-    required?: string[];
-  };
-  annotations?: ToolAnnotations;
-}
-
-export interface ToolResponse {
-  content: Array<{ type: string; text?: string; [k: string]: unknown }>;
-  isError?: boolean;
-  [k: string]: unknown;
-}
-
-export type ToolHandler = (
-  runner: GodotRunner,
-  args: OperationParams,
-) => Promise<ToolResponse> | ToolResponse;
-
-// Parameter mappings between snake_case and camelCase
-// Add new entries whenever a tool surfaces a new compound parameter — the
-// strict converter throws in test env on unmapped keys to catch oversights.
-const parameterMappings: Record<string, string> = {
-  project_path: 'projectPath',
-  scene_path: 'scenePath',
-  root_node_type: 'rootNodeType',
-  parent_node_path: 'parentNodePath',
-  parent_path: 'parentPath',
-  node_type: 'nodeType',
-  node_name: 'nodeName',
-  texture_path: 'texturePath',
-  node_path: 'nodePath',
-  node_paths: 'nodePaths',
-  target_node_path: 'targetNodePath',
-  target_parent_path: 'targetParentPath',
-  new_name: 'newName',
-  output_path: 'outputPath',
-  mesh_item_names: 'meshItemNames',
-  new_path: 'newPath',
-  file_path: 'filePath',
-  script_path: 'scriptPath',
-  response_mode: 'responseMode',
-  preview_max_width: 'previewMaxWidth',
-  preview_max_height: 'previewMaxHeight',
-  bridge_port: 'bridgePort',
-  abort_on_error: 'abortOnError',
-  max_depth: 'maxDepth',
-  changed_only: 'changedOnly',
-  case_sensitive: 'caseSensitive',
-  file_types: 'fileTypes',
-  max_results: 'maxResults',
-};
-
-// Reverse mapping from camelCase to snake_case
-const reverseParameterMappings: Record<string, string> = {};
-for (const [snakeCase, camelCase] of Object.entries(parameterMappings)) {
-  reverseParameterMappings[camelCase] = snakeCase;
-}
-
-export function normalizeParameters(params: OperationParams): OperationParams {
-  if (!params || typeof params !== 'object') {
-    return params;
-  }
-
-  const result: OperationParams = {};
-
-  for (const key in params) {
-    if (Object.prototype.hasOwnProperty.call(params, key)) {
-      let normalizedKey = key;
-
-      if (key.includes('_') && parameterMappings[key]) {
-        normalizedKey = parameterMappings[key];
-      }
-
-      const value = params[key];
-      if (typeof value === 'object' && value !== null && !Array.isArray(value)) {
-        result[normalizedKey] = normalizeParameters(value as OperationParams);
-      } else {
-        result[normalizedKey] = value;
-      }
-    }
-  }
-
-  return result;
-}
-
-function convertCamelToSnakeValue(value: unknown): unknown {
-  if (Array.isArray(value)) {
-    return value.map((entry) => convertCamelToSnakeValue(entry));
-  }
-  if (typeof value === 'object' && value !== null) {
-    return convertCamelToSnakeCase(value as OperationParams);
-  }
-  return value;
-}
-
-export function convertCamelToSnakeCase(params: OperationParams): OperationParams {
-  const result: OperationParams = {};
-  const isTestEnv = process.env.NODE_ENV === 'test' || process.env.VITEST === 'true';
-
-  for (const key in params) {
-    if (Object.prototype.hasOwnProperty.call(params, key)) {
-      const mapped = reverseParameterMappings[key];
-      let snakeKey: string;
-      if (mapped) {
-        snakeKey = mapped;
-      } else if (/[A-Z]/.test(key)) {
-        // Unmapped camelCase key — tolerated in production via regex fallback,
-        // but in tests we throw to catch missing entries in parameterMappings.
-        if (isTestEnv) {
-          throw new Error(
-            `convertCamelToSnakeCase: unmapped camelCase key '${key}'. ` +
-              `Add it to parameterMappings in src/utils/godot-runner.ts so snake/camel conversion stays explicit.`,
-          );
-        }
-        snakeKey = key.replace(/[A-Z]/g, (letter) => `_${letter.toLowerCase()}`);
-      } else {
-        snakeKey = key;
-      }
-      result[snakeKey] = convertCamelToSnakeValue(params[key]) as OperationParams[string];
-    }
-  }
-
-  return result;
-}
-
-/**
- * Check whether a display server (X11 / Wayland) is available on the current
- * platform.  On macOS and Windows the display subsystem is always present;
- * on Linux we probe the standard environment variables.
- */
-export function checkDisplayAvailable(): boolean {
-  if (process.platform !== 'linux') return true;
-  return !!(process.env.DISPLAY || process.env.WAYLAND_DISPLAY);
-}
-
-export function validatePath(path: string): boolean {
-  if (!path || path.includes('..')) {
-    return false;
-  }
-  return true;
-}
-
-/**
- * Stricter check for paths that must stay inside `projectPath`. Rejects `..`
- * (via `validatePath`) and absolute paths that escape the project root.
- * `path.join('/project', '/etc/passwd')` resolves to `/etc/passwd`, so the
- * basic `..`-substring check alone permits absolute-path traversal.
- *
- * Tolerates a leading `res://` (Godot's project-root URI) by stripping it
- * before resolving — autoload entries and resource paths use this prefix.
- */
-export function validateSubPath(projectPath: string, userPath: string): boolean {
-  if (!validatePath(userPath)) return false;
-  const stripped = userPath.startsWith('res://') ? userPath.slice('res://'.length) : userPath;
-  if (!stripped) return false;
-  const projectRoot = resolve(projectPath);
-  const resolved = resolve(projectRoot, stripped);
-  const tail = projectRoot === sep ? sep : projectRoot + sep;
-  return resolved === projectRoot || resolved.startsWith(tail);
-}
-
-/**
- * Validate a Godot scene-tree path (NodePath). Scene-tree paths are a
- * separate namespace from filesystem paths — they address nodes inside
- * a scene, not files on disk, so the project-root containment check
- * in `validateSubPath` does not apply.
- *
- * Rejects empty strings and `..` segments. Accepts both relative
- * (`root/Player`) and absolute (`/root/Player`) Godot forms; the
- * codebase convention is the relative form.
- */
-export function validateNodePath(path: string): boolean {
-  return typeof path === 'string' && path.length > 0 && !path.includes('..');
-}
-
-/**
- * True when `child` resolves to `parent` or a path beneath it. Used by
- * defense-in-depth checks on bridge-returned paths (e.g. screenshot files
- * that must live under `.mcp/screenshots/`).
- */
-export function isUnderDir(parent: string, child: string): boolean {
-  const parentResolved = resolve(parent);
-  const childResolved = resolve(child);
-  return childResolved === parentResolved || childResolved.startsWith(parentResolved + sep);
-}
-
-/**
- * Return `error.message` when `error` is an `Error`, otherwise `'Unknown error'`.
- * Centralizes the catch-block boilerplate so handlers can build error responses
- * without repeating the `instanceof Error` ternary.
- */
-export function getErrorMessage(error: unknown): string {
-  return error instanceof Error ? error.message : 'Unknown error';
-}
-
-/**
- * Build the absolute path to a project's `project.godot` manifest. Use this
- * instead of `join(dir, 'project.godot')` ad hoc.
- */
-export function projectGodotPath(projectDir: string): string {
-  return join(projectDir, 'project.godot');
-}
-
-/**
- * Extract the first [ERROR] message from GDScript stderr output.
- * Falls back to a generic message if no [ERROR] line is found.
- */
-export function extractGdError(stderr: string): string {
-  const errLine = stderr.split('\n').find((l) => l.includes('[ERROR]'));
-  return errLine
-    ? errLine.replace(/.*\[ERROR\]\s*/, '').trim()
-    : 'see get_debug_output for details';
-}
-
-export function createErrorResponse(
-  message: string,
-  possibleSolutions: string[] = [],
-): {
-  content: Array<{ type: 'text'; text: string }>;
-  isError: boolean;
-} {
-  logError(`Error response: ${message}`);
-  if (possibleSolutions.length > 0) {
-    logError(`Possible solutions: ${possibleSolutions.join(', ')}`);
-  }
-
-  const response: {
-    content: Array<{ type: 'text'; text: string }>;
-    isError: boolean;
-  } = {
-    content: [{ type: 'text', text: message }],
-    isError: true,
-  };
-
-  if (possibleSolutions.length > 0) {
-    response.content.push({
-      type: 'text',
-      text: 'Possible solutions:\n- ' + possibleSolutions.join('\n- '),
-    });
-  }
-
-  return response;
-}
-
-// --- Shared validation helpers ---
-
-interface ValidatedProjectArgs {
-  projectPath: string;
-}
-
-interface ValidatedSceneArgs {
-  projectPath: string;
-  scenePath: string;
-}
-
-type ValidationErrorResult = ReturnType<typeof createErrorResponse>;
-
-export function validateProjectArgs(
-  args: OperationParams,
-): ValidatedProjectArgs | ValidationErrorResult {
-  if (!args.projectPath) {
-    return createErrorResponse('projectPath is required', [
-      'Provide a valid path to a Godot project directory',
-    ]);
-  }
-
-  if (!validatePath(args.projectPath as string)) {
-    return createErrorResponse('Invalid project path', [
-      'Provide a valid path without ".." or other potentially unsafe characters',
-    ]);
-  }
-
-  const projectFile = projectGodotPath(args.projectPath as string);
-  if (!existsSync(projectFile)) {
-    return createErrorResponse(`Not a valid Godot project: ${args.projectPath}`, [
-      'Ensure the path points to a directory containing a project.godot file',
-    ]);
-  }
-
-  return { projectPath: args.projectPath as string };
-}
-
-export function validateSceneArgs(
-  args: OperationParams,
-  opts?: { sceneRequired?: boolean },
-): ValidatedSceneArgs | ValidationErrorResult {
-  const projectResult = validateProjectArgs(args);
-  if ('isError' in projectResult) return projectResult;
-
-  const sceneRequired = opts?.sceneRequired !== false;
-
-  if (!args.scenePath) {
-    if (sceneRequired) {
-      return createErrorResponse('scenePath is required', [
-        'Provide the scene file path relative to the project',
-      ]);
-    }
-    return { projectPath: projectResult.projectPath, scenePath: '' };
-  }
-
-  if (!validateSubPath(projectResult.projectPath, args.scenePath as string)) {
-    return createErrorResponse('Invalid scene path', [
-      'Provide a valid relative path without ".." that stays inside the project directory',
-    ]);
-  }
-
-  if (sceneRequired) {
-    const sceneFullPath = join(projectResult.projectPath, args.scenePath as string);
-    if (!existsSync(sceneFullPath)) {
-      return createErrorResponse(`Scene file does not exist: ${args.scenePath}`, [
-        'Ensure the scene path is correct',
-        'Use create_scene to create a new scene first',
-      ]);
-    }
-  }
-
-  return { projectPath: projectResult.projectPath, scenePath: args.scenePath as string };
 }
 
 interface InFlightCommand {
