+++ tests/unit/tool-definitions.test.ts
@@ -5,7 +5,7 @@ import { projectToolDefinitions } from '../../src/tools/project-tools.js';
 import { sceneToolDefinitions } from '../../src/tools/scene-tools.js';
 import { nodeToolDefinitions } from '../../src/tools/node-tools.js';
 import { validateToolDefinitions } from '../../src/tools/validate-tools.js';
-import type { ToolDefinition } from '../../src/utils/godot-runner.js';
+import type { ToolDefinition } from '../../src/mcp.types.js';
 
 const allDefinitions: ToolDefinition[] = [
   ...runtimeToolDefinitions,
